Adobe1A/requirements.txt
pdfminer.six==20221105
langdetect==1.0.9
